web-portal-automator-src/requirements.txt
selenium
gspread
pandas
google-auth
google-auth-oauthlib
colorama
keyboard
chromedriver-autoinstaller
